// src/app/services/chart/model/chart-api.model.ts
export class ChartApiModel {
  chart: {
    backgroundColor: any;
    style: any;
  }
  colors: any;
  credits: {
    enabled: false;
  }
  data: any;
  exporting: any;
  labels: any;
  legend: any;
  plotOptions: any;
  series: Array<any>;
  subtitle: any;
  title: any;
  tooltip: any;
  xAxis: any;
  yAxis: any;
  zAxis: any;
}